Wrote tests for all known good workouts, fixed bug in time basis lap check
The existing logic attempted to check if a TIME basis'd lap was within MAX_TIME_DIFF (5 seconds) of the set time. The aggregate lap logic sort of bypassed this, as the check was done on the individual lap level, but the basis was already set. The fix computes a percentage  of laps that are outside of the MAX_TIME_DIFF and if it's above MAX_TIME_DIFF_RATIO (.25), flips the entire set's basis to DISTANCE.